21 wip

+++ cvis/src/postprocessing/film.h
@@ -12,15 +12,15 @@ struct FilmPass {
 		shader = std::make_unique<gl::Shader>("shaders/uv2d.vert", "shaders/filmShader.frag");
 	}
 
-	void draw(float time, float noiseLevel, float scanlineLevel, float scanlineCount) {
+	void draw(double time, double noiseLevel, double scanlineLevel, double scanlineCount) {
 		glContext->framebufferBack->activate();
 		glClearColor(1.0f, 0.0f, 1.0f, 1.0f);
 		glClear(GL_COLOR_BUFFER_BIT);
 		shader->use();
-		shader->setFloat("seed", time - (long)time);
-		shader->setFloat("noiseLevel", noiseLevel);
-		shader->setFloat("scanlineLevel", scanlineLevel);
-		shader->setFloat("scanlineCount", scanlineCount);
+		shader->setFloat("seed", (float)(time - (long)time));
+		shader->setFloat("noiseLevel", (float)noiseLevel);
+		shader->setFloat("scanlineLevel", (float)scanlineLevel);
+		shader->setFloat("scanlineCount", (float)scanlineCount);
 		glContext->drawFullScreen(glContext->framebufferFront->texture);
 		glContext->swapFramebuffers();
 	}
